pyproject.toml
[project]
name = "trust-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-firestore>=2.20.2",
    "ipykernel>=6.29.5",
    "numpy>=2.2.5",
    "openai>=1.76.2",
    "python-dotenv>=1.1.0",
    "python-engineio>=4.12.0",
    "python-socketio>=5.13.0",
    "reflex>=0.7.9",
    "requests>=2.32.3",
    "toml>=0.10.2",
    "urllib3>=2.4.0",
]
